# Conflicts:
#	gieFrontend/src/main/java/org/giefront/EditDeleteCell.java
#	gieFrontend/src/main/java/org/giefront/EditDeleteCellEnt.java
#	gieFrontend/src/main/java/org/giefront/EntropriseModificationController.java
#	gieFrontend/src/main/java/org/giefront/FournisseurPersoController.java
#	gieFrontend/src/main/java/org/giefront/FournisseurPersoControlleur.java
#	gieFrontend/src/main/java/org/giefront/IEntropriseModificationController.java
#	gieFrontend/src/main/resources/org/Interfaces/fournisseurEntro.fxml
#	gieFrontend/src/main/resources/org/Interfaces/fournisseurPerso.fxml

+++ gieBackend/src/main/java/org/gieback/Controller/EntrepriseController.java
@@ -56,7 +56,7 @@ public class EntrepriseController {
 
     @DELETE
     @Path("DeleteEntreprise/{id}")
-    public Response deleteByid(@PathParam("id") int id) {
+    public Response deleteByid(@PathParam("id") Long id) {
         entrepriseService.deleteByid(id);
         return Response.noContent().build();
     }
